move market mock to utils file
