+++ zrml/parimutuel/src/lib.rs
@@ -24,6 +24,7 @@ extern crate alloc;
 mod benchmarking;
 mod mock;
 mod tests;
+mod utils;
 pub mod weights;
 
 pub use pallet::*;
@@ -469,40 +470,5 @@ mod pallet {
                 MarketType::Scalar(_) => Err(Error::<T>::ScalarMarketsNotAllowed.into()),
             }
         }
-    }
-}
-
-#[cfg(any(feature = "runtime-benchmarks", test))]
-pub(crate) fn market_mock<T>() -> MarketOf<T>
-where
-    T: crate::Config,
-{
-    use frame_support::traits::Get;
-    use sp_runtime::{traits::AccountIdConversion, Perbill};
-    use zeitgeist_primitives::types::{
-        Asset, Deadlines, MarketBonds, MarketCreation, MarketDisputeMechanism, MarketPeriod,
-        MarketStatus, MarketType, ScoringRule,
-    };
-
-    zeitgeist_primitives::types::Market {
-        base_asset: Asset::Ztg,
-        creation: MarketCreation::Permissionless,
-        creator_fee: Perbill::zero(),
-        creator: T::PalletId::get().into_account_truncating(),
-        market_type: MarketType::Scalar(0..=100),
-        dispute_mechanism: Some(MarketDisputeMechanism::Authorized),
-        metadata: Default::default(),
-        oracle: T::PalletId::get().into_account_truncating(),
-        period: MarketPeriod::Block(Default::default()),
-        deadlines: Deadlines {
-            grace_period: 1_u32.into(),
-            oracle_duration: 1_u32.into(),
-            dispute_duration: 1_u32.into(),
-        },
-        report: None,
-        resolved_outcome: None,
-        scoring_rule: ScoringRule::Parimutuel,
-        status: MarketStatus::Active,
-        bonds: MarketBonds::default(),
     }
 }
